Handled Server Rep Case
If the listen server machine's player class overlaps with the weapon sphere, the rep notify function isn't called, so I made a new function that checks whether the pawn IsLocallyControlled(), and if it is, show the widget. Clients are NOT locally controlled, but the listen server is. Listen server machines DO NOT call rep notify functions on themselves